# docs/swagger.yaml
definitions:
  domain.PaymentMethod:
    properties:
      id:
        type: integer
      payment_method:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Glamgrove is an E-commerce platform for purchasing high quality beauty
    products
  title: Go + Gin E-Commerce API Glamgrove
  version: 1.0.0
paths:
...
  /admin/paymentmethod/add:
    post:
      consumes:
      - application/json
      description: Adds a new payment method based on the provided details.
      parameters:
      - description: Payment method details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentMethod'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added payment method
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error while fetching data from user" or "Can't add payment
            method
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add a new payment method
      tags:
      - Payment Methods
  /admin/paymentmethod/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a payment method by its ID.
      parameters:
      - description: Payment method ID to delete
        in: query
        name: methodID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a payment method
      tags:
      - Payment Methods
  /admin/paymentmethod/update:
    put:
      consumes:
      - application/json
      description: Updates an existing payment method.
      parameters:
      - description: Payment method object to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentMethod'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error while getting data or invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a payment method
      tags:
      - Payment Methods
  /admin/paymentmethod/view:
    get:
      consumes:
      - application/json
      description: Retrieves a list of payment methods with pagination support.
      parameters:
      - description: Number of items per page
        in: query
        name: count
        type: integer
      - description: Page number
        in: query
        name: page_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of payment methods
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get payment methods
      tags:
      - Payment Methods
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"